Rename misspelled service python files
