+++ tests/components/color_extractor/test_services.py
