+++ tests/components/music_assistant/test_services.py
@@ -6,15 +6,15 @@ from music_assistant_models.media_items import SearchResults
 import pytest
 from syrupy.assertion import SnapshotAssertion
 
-from homeassistant.components.music_assistant.actions import (
-    SERVICE_GET_LIBRARY,
-    SERVICE_SEARCH,
-)
 from homeassistant.components.music_assistant.const import (
     ATTR_FAVORITE,
     ATTR_MEDIA_TYPE,
     ATTR_SEARCH_NAME,
     DOMAIN,
+)
+from homeassistant.components.music_assistant.services import (
+    SERVICE_GET_LIBRARY,
+    SERVICE_SEARCH,
 )
 from homeassistant.const import ATTR_CONFIG_ENTRY_ID
 from homeassistant.core import HomeAssistant
